requirements.txt
# Multi-Modal RAG Agent Dependencies

# Core Azure SDK
azure-identity>=1.15.0
openai>=1.12.0
azure-storage-blob>=12.19.0

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# PDF processing
pdfminer.six>=20221105

# Optional: For HTTP server mode
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0